# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Aether Guide API
  version: 0.1.0
paths:
...
  /api/v1/location/visual:
    post:
      tags:
      - location
      summary: Locate Visual
      operationId: locate_visual_api_v1_location_visual_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisualLocationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse_LocationResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/location/qr:
    post:
      tags:
      - location
      summary: Locate Qr
      operationId: locate_qr_api_v1_location_qr_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QRLocationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse_LocationResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/location/conversational:
    post:
      tags:
      - location
      summary: Locate Conversational
      operationId: locate_conversational_api_v1_location_conversational_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationalLocationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse_LocationResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/location/fuse:
    post:
      tags:
      - location
      summary: Fuse Location
      operationId: fuse_location_api_v1_location_fuse_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuseLocationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse_LocationResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/location/trail:
    delete:
      tags:
      - location
      summary: Clear Trail
      operationId: clear_trail_api_v1_location_trail_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearTrailRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse_dict_str__bool__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BaseResponse_LocationResult_:
      properties:
        data:
          $ref: '#/components/schemas/LocationResult'
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: ok
        trace_id:
          type: string
          title: Trace Id
      type: object
      required:
      - data
      - trace_id
      title: BaseResponse[LocationResult]
...
    BaseResponse_dict_str__bool__:
      properties:
        data:
          additionalProperties:
            type: boolean
          type: object
          title: Data
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: ok
        trace_id:
          type: string
          title: Trace Id
      type: object
      required:
      - data
      - trace_id
      title: BaseResponse[dict[str, bool]]
...
    ClearTrailRequest:
      properties:
        user_id:
          type: string
          title: User Id
        scenic_id:
          type: string
          title: Scenic Id
      additionalProperties: false
      type: object
      required:
      - user_id
      - scenic_id
      title: ClearTrailRequest
    ConversationalLocationRequest:
      properties:
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Idempotency Key
          description: Client-generated idempotency key for safe retries.
        scenic_id:
          type: string
          title: Scenic Id
        description:
          type: string
          maxLength: 1000
          minLength: 2
          title: Description
      additionalProperties: false
      type: object
      required:
      - scenic_id
      - description
      title: ConversationalLocationRequest
...
    FuseLocationRequest:
      properties:
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Idempotency Key
          description: Client-generated idempotency key for safe retries.
        scenic_id:
          type: string
          title: Scenic Id
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
        inputs:
          items:
            $ref: '#/components/schemas/LocationInput'
          type: array
          minItems: 1
          title: Inputs
      additionalProperties: false
      type: object
      required:
      - scenic_id
      - inputs
      title: FuseLocationRequest
    GeoPoint:
      properties:
        lat:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Lat
        lng:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Lng
        accuracy_m:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Accuracy M
      additionalProperties: false
      type: object
      required:
      - lat
      - lng
      title: GeoPoint
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LocationInput:
      properties:
        source:
          $ref: '#/components/schemas/LocationSource'
        point:
          $ref: '#/components/schemas/GeoPoint'
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
      additionalProperties: false
      type: object
      required:
      - source
      - point
      - confidence
      title: LocationInput
    LocationResult:
      properties:
        status:
          type: string
          title: Status
        scenic_id:
          type: string
          title: Scenic Id
        landmark_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Landmark Id
        point:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
        confidence:
          type: number
          title: Confidence
        follow_up:
          anyOf:
          - type: string
          - type: 'null'
          title: Follow Up
      additionalProperties: false
      type: object
      required:
      - status
      - scenic_id
      - landmark_id
      - point
      - confidence
      title: LocationResult
    LocationSource:
      type: string
      enum:
      - gps
      - vps
      - qr
      - conversational
      - pdr
      title: LocationSource
...
    QRLocationRequest:
      properties:
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Idempotency Key
          description: Client-generated idempotency key for safe retries.
        scenic_id:
          type: string
          title: Scenic Id
        poi_id:
          type: string
          title: Poi Id
        token:
          type: string
          minLength: 8
          title: Token
      additionalProperties: false
      type: object
      required:
      - scenic_id
      - poi_id
      - token
      title: QRLocationRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VisualLocationRequest:
      properties:
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: 'null'
          title: Idempotency Key
          description: Client-generated idempotency key for safe retries.
        scenic_id:
          type: string
          title: Scenic Id
        image_base64:
          type: string
          minLength: 16
          title: Image Base64
        gps_hint:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - scenic_id
      - image_base64
      title: VisualLocationRequest